Added tracker measuring, confirmed tracker/actuators work.
I ran this revision with the laser tracker and actuators powered
on, and confirmed that you could tell the actuators to move and
measure that movement with the laser tracker (well, for one
actuator [port 194], at least -- didn't test others yet). I think
that means this is the first completely functional revision! Yay!
Break out the party hats!